Improve local config
This changes the structure of the local config in $HOME/.config/gofile-upload/config.json
and also how configuration options are passed to the client and API.
Instead of function params, they now take an `options` dict with all possible configs.

The client options are descrbied via GofileUploaderOptions and is based on
GofileUploaderLocalConfigOptions which is the configs that are saved locally.

The CLI/argparse code has also been separated to its own file and now
contains the logic for loading the local config file.

The current logic is that local config is loaded, then overridden with the config
provided by the CLI.
This seems ok at a glance but currently argparse has some sane defaults so local
config would never actually used as they would be overridden by the default cli params which
are not specified. Changes for this will come in a future commit.

All of this is also untested as this is a WIP feature so don't expect stuff to work.